move files Options and SubscriberInfo
