+++ README.md
@@ -11,8 +11,8 @@ API for expert sender service
 use GuzzleHttp\Client;
 use Citilink\ExpertSenderApi\RequestSender;
 use Citilink\ExpertSenderApi\ExpertSenderApi;
-use Citilink\ExpertSenderApi\Model\SubscribersRequest\SubscriberInfo;
-use Citilink\ExpertSenderApi\Model\SubscribersRequest\Options;
+use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\SubscriberInfo;
+use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\Options;
 use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\Identifier;
 use Citilink\ExpertSenderApi\Enum\SubscribersPostRequest\Mode;
