+++ src/Request/SubscribersPostRequest.php
@@ -6,8 +6,8 @@ namespace Citilink\ExpertSenderApi\Request;
 use Citilink\ExpertSenderApi\Enum\HttpMethod;
 use Citilink\ExpertSenderApi\Enum\SubscribersPostRequest\MatchingMode;
 use Citilink\ExpertSenderApi\Exception\ExpertSenderApiException;
-use Citilink\ExpertSenderApi\Model\SubscribersRequest\Options;
-use Citilink\ExpertSenderApi\Model\SubscribersRequest\SubscriberInfo;
+use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\Options;
+use Citilink\ExpertSenderApi\Model\SubscribersPostRequest\SubscriberInfo;
 use Citilink\ExpertSenderApi\RequestInterface;
 use Citilink\ExpertSenderApi\Utils;
 use Webmozart\Assert\Assert;
